requirements.txt
langchain
python-dotenv
langchain_openai
langchain_community
langchain-chroma